docs: add LineChartHasMultiSeries example in TypeScript (.tsx)
